[dawn][native] Spanify SetImmediates.
Also move the common recording code to ProgrammableEncoder.

Bug: 515272358, 528305452, 524406299
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/320295
Commit-Queue: Corentin Wallez <cwallez@chromium.org>

+++ src/dawn/native/CommandBufferStateTracker.h
@@ -77,7 +77,7 @@ class CommandBufferStateTracker {
                         uint64_t size);
     void UnsetVertexBuffer(VertexBufferSlot slot);
     void SetVertexBuffer(VertexBufferSlot slot, uint64_t size);
-    void SetImmediateData(uint32_t offset, uint32_t size);
+    void SetImmediateData(uint32_t offset, size_t size);
     void End();
 
     static constexpr size_t kNumAspects = 6;
